+++ .gitignore
@@ -2,4 +2,3 @@
 .env
 bun.*
 .idea/
-/scratch/
