+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "npx jest",
-    "build": "NODE_ENV=production node src/index",
-    "dev": "NODE_ENV=development npx nodemon src/index",
+    "build": "node src/index",
+    "dev": "npx nodemon src/index",
     "undo:migrate": "npx sequelize-cli db:migrate:undo:all",
     "undo:seed": "npx sequelize-cli db:seed:undo:all",
     "migrate": "npx sequelize-cli db:migrate",
@@ -37,8 +37,6 @@
     "mongoose": "^8.3.2",
     "multer": "^1.4.5-lts.1",
     "mysql2": "^3.9.6",
-    "node-localstorage": "^3.0.5",
-    "qs": "^6.12.1",
     "sequelize": "^6.37.3",
     "supertest": "^6.3.4"
   },
