+++ src/database/seeders/20240504172642-demo-product.js
@@ -20,18 +20,18 @@ module.exports = {
                 min: 10000,
                 max: 30000000,
             });
-            const id = faker.string.uuid();
+            const id = faker.commerce.isbn({
+                separator: "-",
+            });
             const seller =
                 sellers[Math.floor(Math.random() * sellers.length)].dataValues
                     .username;
-            const weight = Math.floor(Math.random() * 20);
             products.push({
                 id: id,
                 name: name,
                 description: description,
                 price: price,
                 seller: seller,
-                weight: weight
             });
         }
         return queryInterface.bulkInsert("products", products);
