+++ src/database/seeders/20240504180302-demo-transaction.js
@@ -3,7 +3,18 @@ const { fakerID_ID: faker } = require("@faker-js/faker");
 /** @type {import('sequelize-cli').Migration} */
 const { User } = require("../../model");
 const { DateTime } = require("luxon");
-const { addressGenerator } = require("../../utility/address");
+
+const addressGenerator = () => {
+    const country = "Indonesia";
+    const city = faker.location.city();
+    const state = faker.location.state();
+    const address = faker.location.streetAddress();
+    const zipCode = faker.location.zipCode();
+
+    const buildingNumber = faker.location.buildingNumber();
+
+    return `${address} No ${buildingNumber}, ${zipCode}, ${city}, ${state}, ${country}`;
+};
 
 module.exports = {
     async up(queryInterface, Sequelize) {
@@ -16,19 +27,25 @@ module.exports = {
         });
         for (let i = 0; i < 20; i++) {
             const id = faker.string.uuid();
+            const origin = addressGenerator();
             const destination = addressGenerator();
+            const deliveryPrice = faker.commerce.price({
+                min: 20000,
+                max: 200000,
+            });
             const buyer =
                 buyers[Math.floor(Math.random() * buyers.length)].dataValues
                     .username;
-            const deliveryPrice = faker.commerce.price({
-                min: 20000,
-                max: 50000
+            const arrival = faker.date.future({
+                refDate: DateTime.now(),
             });
             transactions.push({
                 id: id,
+                origin: origin,
                 destination: destination,
+                delivery_price: deliveryPrice,
                 buyer: buyer,
-                delivery_price: deliveryPrice
+                arrival_date: arrival,
             });
         }
         return queryInterface.bulkInsert("transactions", transactions);
