+++ src/model/models/Transaction.js
@@ -26,6 +26,10 @@ module.exports = (sequelize, DataTypes) => {
                 type: DataTypes.STRING(255),
                 allowNull: false,
             },
+            origin: {
+                type: DataTypes.STRING(255),
+                allowNull: false,
+            },
             destination: {
                 type: DataTypes.STRING(255),
                 allowNull: false,
@@ -33,6 +37,10 @@ module.exports = (sequelize, DataTypes) => {
             deliveryPrice: {
                 type: DataTypes.DECIMAL.UNSIGNED,
                 allowNull: false,
+            },
+            arrivalDate: {
+                type: Sequelize.DATE,
+                allowNull: true,
             },
         },
         {
