+++ src/model/models/TransactionDetail.js
@@ -1,16 +1,8 @@
 "use strict";
-const { Model, Sequelize} = require("sequelize");
+const { Model } = require("sequelize");
 module.exports = (sequelize, DataTypes) => {
     class TransactionDetail extends Model {
-        static associate(models) {
-            TransactionDetail.belongsTo(models.Product, {
-                foreignKey: 'product',
-            });
-
-            TransactionDetail.belongsTo(models.Transaction, {
-                foreignKey: 'transaction',
-            });
-        }
+        static associate(models) {}
     }
     TransactionDetail.init(
         {
@@ -36,17 +28,10 @@ module.exports = (sequelize, DataTypes) => {
                 type: DataTypes.DECIMAL.UNSIGNED,
                 allowNull: false,
             },
-            deliveryDate: {
-                type: Sequelize.DATE,
-                allowNull: true,
-            },
         },
         {
             sequelize,
             modelName: "TransactionDetail",
-            tableName: "transaction_details",
-            underscored: true,
-            timestamps: false
         }
     );
     return TransactionDetail;
