--- src/routes/api/roles.js
@@ -1,7 +1,0 @@
-const { getAdminRole } = require("../../controller/api/getAdmin")
-const { checkAccessToken, checkRole } = require("../../middleware/token")
-
-module.exports = (expressApp) => ({
-    middleware: [checkAccessToken, checkRole("ADM")],
-    get: getAdminRole
-})
