--- src/routes/api/tiers.js
@@ -1,7 +1,0 @@
-const { getAdminTier } = require("../../controller/api/getAdmin")
-const { checkAccessToken, checkRole } = require("../../middleware/token")
-
-module.exports = (expressApp) => ({
-    middleware: [checkAccessToken, checkRole("ADM")],
-    get: getAdminTier
-})
