--- src/routes/api/transactions.js
@@ -1,7 +1,0 @@
-const { getMyTransaction } = require("../../controller/api/transaction");
-const { checkAccessToken, checkMoreRole } = require("../../middleware/token");
-
-module.exports = (expressApp) => ({
-    middleware: [checkAccessToken, checkMoreRole("USR,SLR")],
-    get: getMyTransaction
-});
