--- src/routes/api/users/cart.js
@@ -1,7 +1,0 @@
-const { addToCart, getCart } = require("../../../controller/api/user");
-const { checkAccessToken, checkRole } = require("../../../middleware/token");
-module.exports = (expressApp) => ({
-    middleware: [checkAccessToken, checkRole("USR")],
-    get: getCart,
-    post: addToCart
-});
