--- src/routes/api/users/checkout.js
@@ -1,6 +1,0 @@
-const { checkout } = require("../../../controller/api/user");
-const { checkAccessToken } = require("../../../middleware/token");
-module.exports = (expressApp) => ({
-    middleware: checkAccessToken,
-    post: checkout
-})
