--- src/routes/api/users/{id}.js
@@ -1,8 +1,0 @@
-const { updateData } = require("../../../controller/api/updateUser");
-const { checkPermission } = require("../../../middleware/access");
-const { checkAccessToken } = require("../../../middleware/token");
-
-module.exports = (expressApp) => ({
-    middleware: [checkAccessToken, checkPermission],
-    put: updateData
-});
