+++ src/routes/api/users/{id}/cart.js
@@ -1,0 +1,5 @@
+module.exports = (expressApp) => ({
+    get: (request, response) => {
+        response.status(200).send("Hello, Route").end();
+    },
+});
