--- src/routes/auth/logout.js
@@ -1,4 +1,0 @@
-const { logoutUser } = require("../../controller/api/auth");
-module.exports = (expressApp) => ({
-    post: logoutUser,
-});
