--- src/service/subscription.js
@@ -1,8 +1,0 @@
-const { Subscription } = require("../model");
-const getSubscription = async (id) => {
-    return (await Subscription.findByPk(id))
-}
-
-module.exports = {
-    getSubscription
-}
