+++ src/service/token.js
@@ -1,6 +1,5 @@
 const { ACCESS_SECRET_KEY, REFRESH_SECRET_KEY } = require("../config");
 const jwt = require("jsonwebtoken");
-const util = require('util');
 
 const createToken = (data, type = "ACCESS", expiresIn = "15m") => {
     const secret = type === "ACCESS" ? ACCESS_SECRET_KEY : REFRESH_SECRET_KEY;
